Add functionality for colour deletion
Now right-clicking on a colour opens a context menu which provides the
option to delete the colour. All colour frame events are rebound on
deletion.

Note: the bug reported in the previous commit message was PEBKAC; was
simply always selecting white from the colour picker because saturation
was full when going in with white as the starting colour.