Isolated the problem pre-sorted state
Although we take all possible 'productive' moves (not into empty tubes)
before any moves into empty tubes, there are situations where we need to
move some liquid into an empty tube to unblock progress.

In these situations, *all other moves into empty tubes can block the
useful move*. Since we proceed left-to-right, if there are single-colour
tubes to the left of the tube that we need to move into the empty space,
those tubes will be shifted in and out of the empty space in all
possible combinations before we make the move we actually need to make.
In DFS, these will all feature in the resulting generated solution. The
factorial number of combinations means this can slow down larger
solutions significantly, as seen on level 135.